logtool. and lots of things in the main firmware.
better integration into main
fixed scheduler and timestamping
simplified the api there
ditched integer 64-bit storage types (not needed. 32 bits is plenty for everything except current time - but that's handled in RTC, everything else is deltas)
web remote!

+++ main/rf_433.h
@@ -16,7 +16,7 @@ int64_t recieveKeycode();
 esp_err_t rf_433_init();
 esp_err_t rf_433_stop();
 
-void rf_433_set_keycode(uint8_t index, int64_t code);
+void rf_433_set_keycode(uint8_t index, uint32_t code);
 
 /*
 int8_t rf_433_get_keycode();
@@ -29,8 +29,8 @@ void rf_433_cancel_learn_keycode();
 void rf_433_disable_controls();
 void rf_433_enable_controls();
 
-int64_t rf_433_get_temp_keycode(uint8_t index);
-void rf_433_set_temp_keycode(uint8_t index, int64_t code);
+int32_t rf_433_get_temp_keycode(uint8_t index);
+void rf_433_set_temp_keycode(uint8_t index, uint32_t code);
 void rf_433_clear_temp_keycodes();
 
 #endif
